+++ serde_components/deserializers/toml/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .toml_reader import tomllib
